# GH-130328: pasting in new REPL is slow on Windows