--- a/Misc/NEWS.d/next/Library/2025-04-24-18-07-49.gh-issue-130328.z7CN8z.rst
+++ b/Misc/NEWS.d/next/Library/2025-04-24-18-07-49.gh-issue-130328.z7CN8z.rst
@@ -1,0 +1,1 @@
+Speedup pasting in ``PyREPL`` on Windows. Fix by Chris Eibl.