@chris-eibl 2025-04-24
Yeah, in case of the legacy console pasting of the 58 lines in the OP takes 15 seconds, and this brings it down to 1.8. Most people will hopefully use  a virtual terminal, where the situation is much less severe. After all, the new REPL can be disabled using https://docs.python.org/3/using/cmdline.html#envvar-PYTHON_BASIC_REPL.