@Fidget-Spinner 2025-04-29
@Yhg1s seeking permission to backport this to 3.13.

According to Chris, it takes 15 seconds to paste 58 lines in the new repl in 3.13. Do you think that's slow enough to be worth backporting?